Move utils package under test
Since the `utils` package is now used by both the integration and the E2E test,
they should be placed in the package structure to represent this.

Move `test/integration/utils` to `test` and update references.
