+++ test/e2e/validators_test.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/shipwright-io/build/pkg/apis"
 	buildv1alpha1 "github.com/shipwright-io/build/pkg/apis/build/v1alpha1"
-	"github.com/shipwright-io/build/test/integration/utils"
+	"github.com/shipwright-io/build/test/utils"
 )
 
 const (
